pyproject.toml
[tool.poetry]
name = "jobberwocky"
version = "0.1.0"
description = "service for job opportunities"
authors = ["Fernandez Gonzalo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
requests = "^2.31.0"
pydantic-settings = "^2.0.2"
httpx = "^0.24.1"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
ruff = "^0.0.282"
pytest = "^7.4.0"
ipython = "^8.14.0"
pytest-httpx = "^0.23.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
